test(integration): consolidate the merge-recompute docker suites into one stack
Each docker-integration test class boots its own full stack (the compose fixture is class-scoped), so several near-identical merge-recompute suites multiplied that fixed boot cost and pushed the suite past the 30-minute cap.

Merge the single-hop, multi-level chain, and delete-read-peer merge/rebase recompute scenarios into one TestMergeRecompute class that boots the stack once. Delete test_merge_derived_value_refresh.py: its main-only node refreshed after a template-change merge is already covered by test_computed_attributes.py::test_merge_triggers_scoped_recompute.
